refactor: error code 리팩토링 (#25)
